Add support to NameFieldResolver for resolving properties of value types.
Currently, an exception is thrown when attempting to resolve a property
of a value type as the signature of the delegate created by
DelegateHelper does not match that of the target method; this is because
the type of the hidden this parameter is actually TTarget& (i.e. IsByRef).

To eliminate the need to explicitly handle ref-types / value-types, we
delegate to System.Linq.Expressions for the heavy lifting.